add checkpoint

+++ fragment_tools/make_fragments.pl
@@ -123,6 +123,7 @@ $options{psipredfile} = "";
 $options{samfile}     = "";
 $options{porterfile}  = "";
 $options{pssm}        = "";
+$options{checkpoint}  = "";
 $options{psipred}     = 1;         # use psipred by default
 $options{porter}      = 0;         # skip porter by default
 $options{sam}         = 0;         # skip sam by default
@@ -454,12 +455,13 @@ if ($SPARKS) {
 }
 
 # if PSSM file provided, skip running blast
-if ( &nonempty_file_exists("$options{pssm}") ) {
+if ( &nonempty_file_exists("$options{pssm}") && &nonempty_file_exists("$options{checkpoint}")) ) {
   system("cp $options{pssm} $options{runid}.pssm");
+  system("cp $options{checkpoint} $options{runid}.check");
 }
 
 # run blast
-unless ( &nonempty_file_exists("$options{runid}.check") || &nonempty_file_exists("$options{pssm}") ) {
+unless ( &nonempty_file_exists("$options{runid}.check") && &nonempty_file_exists("$options{pssm}") ) {
     print_debug("Running psiblast for sequence profile");
     print_debug("Using nr: $NR");
     if (
@@ -1240,6 +1242,7 @@ sub getCommandLineOptions {
 		\t-nofrags specify to not make fragments but run everything else
 		\t-old_name_format  use old name format e.g. aa1tum_05.200_v1_3
     \t-pssm  path to pssm file
+    \t-checkpoint  path to checkpoint file
 		\t<fasta file>
 	};
     $usage = "$usage\n\n" . join ' ', ( 'Version:', VERSION, "\n" );
@@ -1257,7 +1260,8 @@ sub getCommandLineOptions {
         "n_frags=i",        "n_candidates=i",
         "add_vall_files=s", "use_vall_files=s",
         "torsion_bin=s",    "exclude_homologs_by_pdb_date=s",
-        "old_name_format!", "add_pdbs_to_vall=s", "pssm=s"
+        "old_name_format!", "add_pdbs_to_vall=s",
+        "pssm=s", "checkpoint=s"
     );
 
     if ( scalar(@ARGV) != 1 ) {
